requirements.txt
aiogram==2.25.2
aiohttp==3.8.6       
aiosignal==1.3.2     
asgiref==3.8.1       
async-timeout==4.0.3 
attrs==25.1.0        
autobahn==24.4.2     
Automat==24.8.1      
Babel==2.9.1
certifi==2025.1.31   
cffi==1.17.1
channels==4.2.0      
channels_redis==4.2.1
charset-normalizer==3.4.1
constantly==23.10.4
cryptography==44.0.1
daphne==4.1.2
Django==5.1.6
django-crispy-forms==2.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
frozenlist==1.5.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
magic-filter==1.0.12
msgpack==1.1.0
multidict==6.1.0
packaging==24.2
pillow==11.1.0
propcache==0.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.0.0
pytz==2025.1
redis==5.2.1
service-identity==24.2.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
whitenoise==6.9.0
yarl==1.18.3
zope.interface==7.2